requirements.txt
piexif==1.1.3
python-dotenv==0.19.2
requests==2.32.3